Extract directory filtering partial
